requirements.txt
Flask==3.0.0
Jinja2==3.1.2
Werkzeug==3.0.0
PyYAML==3.0.0
google-genai>=0.5.0
